service_graphql_tourisme/app/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
sqlalchemy==2.0.23
strawberry-graphql[fastapi]==0.216.0
